core: ltc: move dh wrappers to separate file
Moves the DH wrappers in tee_ltc_provider.c to its own file, dh.c.

Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
