+++ core/lib/libtomcrypt/src/tee_ltc_provider.c
@@ -125,109 +125,6 @@ static void tee_ltc_reg_algs(void)
 	register_prng(&prng_mpa_desc);
 }
 
-
-/******************************************************************************
- * Asymmetric algorithms
- ******************************************************************************/
-
-#if defined(_CFG_CRYPTO_WITH_ACIPHER)
-
-static bool bn_alloc_max(struct bignum **s)
-{
-	*s = crypto_bignum_allocate(CFG_CORE_BIGNUM_MAX_BITS);
-
-	return *s;
-}
-
-static TEE_Result __maybe_unused convert_ltc_verify_status(int ltc_res,
-							   int ltc_stat)
-{
-	switch (ltc_res) {
-	case CRYPT_OK:
-		if (ltc_stat == 1)
-			return TEE_SUCCESS;
-		else
-			return TEE_ERROR_SIGNATURE_INVALID;
-	case CRYPT_INVALID_PACKET:
-		return TEE_ERROR_SIGNATURE_INVALID;
-	default:
-		return TEE_ERROR_GENERIC;
-	}
-}
-
-#if defined(CFG_CRYPTO_DH)
-
-TEE_Result crypto_acipher_alloc_dh_keypair(struct dh_keypair *s,
-					   size_t key_size_bits __unused)
-{
-	memset(s, 0, sizeof(*s));
-	if (!bn_alloc_max(&s->g)) {
-		return TEE_ERROR_OUT_OF_MEMORY;
-	}
-
-	if (!bn_alloc_max(&s->p))
-		goto err;
-	if (!bn_alloc_max(&s->y))
-		goto err;
-	if (!bn_alloc_max(&s->x))
-		goto err;
-	if (!bn_alloc_max(&s->q))
-		goto err;
-	return TEE_SUCCESS;
-err:
-	crypto_bignum_free(s->g);
-	crypto_bignum_free(s->p);
-	crypto_bignum_free(s->y);
-	crypto_bignum_free(s->x);
-	return TEE_ERROR_OUT_OF_MEMORY;
-}
-
-TEE_Result crypto_acipher_gen_dh_key(struct dh_keypair *key, struct bignum *q,
-				     size_t xbits)
-{
-	TEE_Result res;
-	dh_key ltc_tmp_key;
-	int ltc_res;
-
-	/* Generate the DH key */
-	ltc_tmp_key.g = key->g;
-	ltc_tmp_key.p = key->p;
-	ltc_res = dh_make_key(NULL, find_prng("prng_mpa"), q, xbits,
-			      &ltc_tmp_key);
-	if (ltc_res != CRYPT_OK) {
-		res = TEE_ERROR_BAD_PARAMETERS;
-	} else {
-		ltc_mp.copy(ltc_tmp_key.y,  key->y);
-		ltc_mp.copy(ltc_tmp_key.x,  key->x);
-
-		/* Free the tempory key */
-		dh_free(&ltc_tmp_key);
-		res = TEE_SUCCESS;
-	}
-	return res;
-}
-
-TEE_Result crypto_acipher_dh_shared_secret(struct dh_keypair *private_key,
-					   struct bignum *public_key,
-					   struct bignum *secret)
-{
-	int err;
-	dh_key pk = {
-		.type = PK_PRIVATE,
-		.g = private_key->g,
-		.p = private_key->p,
-		.y = private_key->y,
-		.x = private_key->x
-	};
-
-	err = dh_shared_secret(&pk, public_key, secret);
-	return ((err == CRYPT_OK) ? TEE_SUCCESS : TEE_ERROR_BAD_PARAMETERS);
-}
-
-#endif /* CFG_CRYPTO_DH */
-#endif /* _CFG_CRYPTO_WITH_ACIPHER */
-
-
 TEE_Result crypto_init(void)
 {
 	init_mp_tomcrypt();
